third_party/WebKit/LayoutTests/editing/deleting/merge-unrendered-space.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                'foo<br>',
                '|bar  ',
                '<div>baz</div>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                'foo|bar  ',
                '<div>baz</div>',
            '</div>',
        ].join('')),
        'Delete BR keeps trailing whitespaces');
</script>
